Cleaning up BukkitCommand wrapping, removung unnecessary code
- We just update the BukkitCommand wrapping to be consistent with other parts of the project.
- Removing the configuration loading function that was specific to Bukkit as we now use a global configuration API within Common
